dts/freescale/meticulous: Enable st7701 panel in var symphony
Now that we have all components ready to connect to the
custom st7701 panel we are adding to it the imx8mn-var-som-symphony.dts.
In the future we should for the dts for our own version - however the
dependencies on the imx8mn-var-som.dtsi include device tree limits
our ability to just do a clean overwrite.
To no create one huge dts in one step we continue to work the the
symphony dts for the moment as we are actually testing on the symphony
board.

We are using the wdog1 pin to dim the backlight which means we are unable
to use the wdog1 on the symphony board.

Signed-off-by: Mimoja <git@mimoja.de>