sec_mipi_dsim: prepare panel after clock enablement
The downstream 5.4 and 5.15 sec_mipi_dsim by NXP is trying
to connect to the panel before the internal clocks are
set up. While this is fine with some panels which don't need
a init sequence, for those that do it means that the init
sequence can never be sent. This leads to timeouts and retries
and ultimately to failing panel initialization.
To prevent this we enable the clock before we prepare the panel.

Upstream-status: Only applies to 5.4 kernel - upstream uses its
    own imx8mn DSI driver without the issue

Signed-off-by: Mimoja <git@mimoja.de>