dts/freescale/meticulous: Add binary and pwm based backlight
The just final Fika boards put the backlight to the I2C4
pins and use them as GPIO5::20 and GPIO5::21. Frustratingly the st7701
panels power supply is also connected to those pins so that we cannot
PWM the signals with current fika boards. Additionally the display
needs to be reset after power-off.
While it is not the best style we are therefore ifdef guarding the
backlight control so that we don't accidentally break workflows with
older board revisions.

Signed-off-by: Mimoja <git@mimoja.de>