arch/arm64/config/meticulous: Disable RNDIS protocol everywhere
While it used to have a point in the times of Windows XP it is
no longer usefull today and inherently insecure.
Disable it and see what breaks, following upstreams descisions to
get rid of it. While upstream hasn't yet fully commited to this,
we dont want to build on top of a broken component anyways.

https://git.kernel.org/pub/scm/linux/kernel/git/gregkh/usb.git/log/?h=rndis-removal

Signed-off-by: Mimoja <mimoja@mimoja.de>